Refactor Like component to own file
